+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.11.5 (Upcoming Release)
 
+- Initial implementation of the robust hat matrix regression estimator
 
 # v0.11.4